added screen drivers, scrolling is BROKEN

+++ build.sh
@@ -11,17 +11,19 @@ nasm kernel_ent.asm -f elf -o ../$DIRECTORY/kernel_ent.o
 
 cd ../$DIRECTORY
 
-gcc -Wall -m32 -ffreestanding -c ../../kernel/kernel.c ../../kernel/busio.c
-ld -m elf_i386 -o kernel.bin -Ttext 0x1000 kernel_ent.o busio.o kernel.o --oformat binary
+gcc -Wall -m32 -ffreestanding -c ../../kernel/kernel.c ../../kernel/busio.c ../../kernel/utils/memory.c ../../drivers/screen/screen.c ../../drivers/screen/text.c
+ld -m elf_i386 -o kernel.bin -Ttext 0x1000 kernel_ent.o busio.o memory.o screen.o text.o kernel.o --oformat binary
 
 cd ../../
 cat $DIRECTORY/boot_sect.bin $DIRECTORY/kernel.bin > os-image.bin
 
-echo "--- Build Complete ---"
-echo "launching bochs..."
-bochs -q
-echo "Done, cleaning up..."
-rm -rf release
-find ./ -name "*.bin" | xargs rm -Rf
-find ./ -name "*.o" | xargs rm -Rf
-echo "Cleaned up."
+if [ "$1" != "-nr" ]; then
+    echo "--- Build Complete ---"
+    echo "launching bochs..."
+    bochs -q
+    echo "Done, cleaning up..."
+    rm -rf release
+    find ./ -name "*.bin" | xargs rm -Rf
+    find ./ -name "*.o" | xargs rm -Rf
+    echo "Cleaned up."
+fi
